Don't allow invalid groups

+++ server/fishtest/views.py
@@ -457,15 +457,18 @@ def user(request):
 
         if request.has_permission("administrate"):
             new_group = request.params.get("group")
-            if (new_group not in user_data["groups"]) and (len(new_group) != len(user_data["groups"])):
-                request.actiondb.change_group(
-                    request.authenticated_userid,
-                    {"user": user_name, "before": user_data["groups"][0], "after": new_group},
-                )
-                request.userdb.remove_user_groups(user_name)
-                if len(new_group) > 0:
-                    request.userdb.add_user_group(user_name, new_group)
-                request.session.flash("Group changed to " + user_data["groups"][0])
+            if new_group == "" or new_group == "group:approvers" or new_group == "group:moderators":
+                if (new_group not in user_data["groups"]) and (len(new_group) != len(user_data["groups"])):
+                    request.actiondb.change_group(
+                        request.authenticated_userid,
+                        {"user": user_name, "before": user_data["groups"][0], "after": new_group},
+                    )
+                    request.userdb.remove_user_groups(user_name)
+                    if len(new_group) > 0:
+                        request.userdb.add_user_group(user_name, new_group)
+                    request.session.flash("Group changed to " + user_data["groups"][0])
+            else:
+                request.session.flash("Invalid group", "error")
         request.userdb.save_user(user_data)
 
     userc = request.userdb.user_cache.find_one({"username": user_name})
